requirements.txt
os
numpy
python-telegram-bot
nest_asyncio
